% Generated by roxygen2 (4.0.2): do not edit by hand
\name{latest.sightings}
\alias{latest.sightings}
\title{Find latest UFO sightings}
\usage{
latest.sightings(area = "BC", return.rows = 5)
}
\arguments{
\item{area}{Two uppercase letters that specifies the area to search}

\item{return.rows}{The number of most recent sightings to return}
}
\description{
This function returns the most recent sightings of Unidentified Flying Objects
in the area (i.e., state, province, territory) specified.
}
\examples{
latest.sightings()
latest.sightings(area='BC')
latest.sightings(area='NC', return.rows = 10)
latest.sightings(area='MA', 3)
}
\keyword{misc}

