% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sightings.by.shape}
\alias{sightings.by.shape}
\title{Visualize/plot UFO sightings by shape}
\usage{
sightings.by.shape(area = "BC", print = "plot")
}
\arguments{
\item{area}{Two letters that specifies the area to search (case is not important)}

\item{print}{The type of object to return, either a ggplot or table}
}
\value{
Either a barplot (generated by ggplot2)
or a table of frequencies
}
\description{
This function returns number of reported sightings of UFOs
in the area (i.e., state, province, territory) grouped by shape, in either
a table or barplot, from \url{http://www.nuforc.org/webreports.html}.
}
\examples{
sightings.by.shape(print = 'plot')
sightings.by.shape(area = 'NY', print = 'plot')
sightings.by.shape(area = 'WA', print = 'table')
}
\seealso{
\code{\link{latest.sightings}}
}
\keyword{misc}

