% Generated by roxygen2 (4.0.2): do not edit by hand
\name{latest.sightings}
\alias{latest.sightings}
\title{Is it a bird? Is it a plane? It's a UFO!}
\usage{
latest.sightings(area = "BC", return.rows = 5, ...)
}
\arguments{
\item{area}{A two-letter term in uppercase that specifies the area to search}

\item{return.rows}{The number of most recent sightings to return}
}
\description{
This function returns the latest sightings of Unidentified Flying Objects
in the area (i.e., state, province, territory) specified.
}
\examples{
latest.sightings()
latest.sightings(area='BC')
latest.sightings(area='NC',10)
latest.sightings(area='MA',3)
}
\keyword{misc}

