% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sightings.by.shape}
\alias{sightings.by.shape}
\title{Visualize/plot UFO sightings by shape}
\usage{
sightings.by.shape(area = "BC", print = "plot")
}
\arguments{
\item{area}{Two uppercase letters that specifies the area to search}

\item{print}{The type of object to return, either a ggplot or table}
}
\description{
This function returns reported sightings of UFOs
in the area (i.e., state, province, territory) specified by shape, in either
a table or barplot.
}
\examples{
sightings.by.shape()
sightings.by.shape(area = 'NY', print = 'plot')
sightings.by.shape(area = 'WA', print = 'table')
}
\keyword{misc}

