% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{ufo}
\alias{ufo}
\alias{ufo-package}
\title{R package: UFO}
\description{
Find the latest reported UFO sightings in different states, provinces and territories across Canda and the USA
}
\details{
The functions in the UFO package help you find the most recent sightings by area and
the types (i.e., shapes) of UFO's reported
}

